+++ db/migrations/postgresql/2018-04-27-192442_nut/up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE votes (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   resource_type VARCHAR(255) NOT NULL,
-  resource_id BIGINT NOT NULL,
+  resource_id INT NOT NULL,
   point INT NOT NULL DEFAULT 0,
   created_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT NOW(), updated_at TIMESTAMP WITH TIME ZONE NOT NULL);
 
@@ -10,14 +10,14 @@ CREATE UNIQUE INDEX idx_votes_resources ON votes (resource_type, resource_id);
 CREATE INDEX idx_votes_resource_type ON votes (resource_type);
 
 CREATE TABLE attachments (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   title VARCHAR(255) NOT NULL,
   url VARCHAR(255) NOT NULL,
   length INT NOT NULL,
   media_type VARCHAR(32) NOT NULL,
   resource_type VARCHAR(255) NOT NULL,
-  resource_id BIGINT NOT NULL,
-  user_id BIGINT NOT NULL,
+  resource_id INT NOT NULL,
+  user_id INT NOT NULL,
   created_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT NOW(), updated_at TIMESTAMP WITH TIME ZONE NOT NULL);
 
 CREATE UNIQUE INDEX idx_attachments_url ON attachments (url);
@@ -29,13 +29,13 @@ CREATE INDEX idx_attachments_resource_type ON attachments (resource_type);
 CREATE INDEX idx_attachments_media_type ON attachments (media_type);
 
 CREATE TABLE leave_words (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   body TEXT NOT NULL,
   type VARCHAR(8) NOT NULL DEFAULT 'markdown',
   created_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT NOW());
 
 CREATE TABLE links (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   href VARCHAR(255) NOT NULL,
   label VARCHAR(255) NOT NULL,
   loc VARCHAR(16) NOT NULL,
@@ -49,7 +49,7 @@ CREATE INDEX idx_links_loc_lang ON links (loc, lang);
 CREATE INDEX idx_links_lang ON links (lang);
 
 CREATE TABLE cards (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   title VARCHAR(255) NOT NULL,
   summary TEXT NOT NULL,
   type VARCHAR(8) NOT NULL DEFAULT 'html',
@@ -66,7 +66,7 @@ CREATE INDEX idx_cards_loc_lang ON cards (loc, lang);
 CREATE INDEX idx_cards_lang ON cards (lang);
 
 CREATE TABLE friend_links (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   title VARCHAR(255) NOT NULL,
   home VARCHAR(255) NOT NULL,
   logo VARCHAR(255) NOT NULL,
