+++ db/migrations/postgresql/2018-04-27-193000_todo/up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE todo_projects (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   title VARCHAR(255) NOT NULL,
   description TEXT,
   created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT now(),
@@ -7,7 +7,7 @@ CREATE TABLE todo_projects (
 );
 
 CREATE TABLE todo_tasks (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   title VARCHAR(255) NOT NULL,
   description TEXT,
   status VARCHAR(8) NOT NULL,
@@ -18,11 +18,11 @@ CREATE TABLE todo_tasks (
 CREATE INDEX todo_tasks_status ON todo_tasks (status);
 
 CREATE TABLE todo_logs (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   message VARCHAR(500) NOT NULL,
   "begin" TIMESTAMP WITHOUT TIME ZONE NOT NULL,
   "end" TIMESTAMP WITHOUT TIME ZONE NOT NULL,
-  task_id BIGINT NOT NULL,
-  member_id BIGINT NOT NULL,
+  task_id INT NOT NULL,
+  member_id INT NOT NULL,
   created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT now()
 );
